Replace 1 useEffect code from PaginationNotif component to PaginationButtonChevron component
